// src/app/model/producto.model.ts
export class productoModel{
    id_producto: number=0;
    fk_categoria: string='';
    nombre: string='';
    costo: string='';
    unidades: string='';
    fecha_creacion: string='';
    usuario_creacion: string='';
    fecha_mod: string='';
    usuario_mod: string='';
    estado: boolean=false;

}